Refactoring Reducer, Store & Action as separate files so that the code looks clean and clear to understand.
