+++ hello-redux/src/app/app.action.ts
@@ -1,8 +1,3 @@
-export const ADD_TODO = 'ADD_TODO';
-export const TOGGLE_TODO = 'TOGGLE_TODO';
-export const REMOVE_TODO = 'REMOVE_TODO';
-export const REMOVE_ALL_TODOS = 'REMOVE_ALL_TODOS';
-
 export const TODOACTIONS = {
     ADD_TODO: 'ADD_TODO',
     TOGGLE_TODO: 'TOGGLE_TODO',
